// amplify/graphql/API.ts
export type Connection = {
  __typename: "Connection",
  connectionId: string,
  createdAt: string,
  updatedAt: string,
  user?: User | null,
  userId: string,
};

export type User = {
  __typename: "User",
  aboutMe?: string | null,
  age?: number | null,
  connectionReceived?: ModelConnectionReceivedConnection | null,
  connectionRequests?: ModelConnectionRequestConnection | null,
  connections?: ModelConnectionConnection | null,
  createdAt: string,
  email: string,
  fullName?: string | null,
  gender?: string | null,
  id: string,
  interests?: Array< string | null > | null,
  profilePictureUrl?: string | null,
  race?: string | null,
  rooms?: ModelRoomUserConnection | null,
  spokenLanguage?: string | null,
  status?: string | null,
  updatedAt: string,
};

export type ModelConnectionReceivedConnection = {
  __typename: "ModelConnectionReceivedConnection",
  items:  Array<ConnectionReceived | null >,
  nextToken?: string | null,
};

export type ConnectionReceived = {
  __typename: "ConnectionReceived",
  createdAt: string,
  receiver?: User | null,
  receiverId: string,
  senderId: string,
  status?: string | null,
  updatedAt: string,
};
// ...
export type DropdownList = {
  __typename: "DropdownList",
  createdAt: string,
  id: string,
  name: string,
  options?: Array< string | null > | null,
  updatedAt: string,
};
// ...
export type GetConnectionQueryVariables = {
  connectionId: string,
  userId: string,
};

export type GetConnectionQuery = {
  getConnection?:  {
    __typename: "Connection",
    connectionId: string,
    createdAt: string,
    updatedAt: string,
    user?:  {
      __typename: "User",
      aboutMe?: string | null,
      age?: number | null,
      createdAt: string,
      email: string,
      fullName?: string | null,
      gender?: string | null,
      id: string,
      interests?: Array< string | null > | null,
      profilePictureUrl?: string | null,
      race?: string | null,
      spokenLanguage?: string | null,
      status?: string | null,
      updatedAt: string,
    } | null,
    userId: string,
  } | null,
};

export type GetConnectionReceivedQueryVariables = {
  receiverId: string,
  senderId: string,
};

export type GetConnectionReceivedQuery = {
  getConnectionReceived?:  {
    __typename: "ConnectionReceived",
    createdAt: string,
    receiver?:  {
      __typename: "User",
      aboutMe?: string | null,
      age?: number | null,
      createdAt: string,
      email: string,
      fullName?: string | null,
      gender?: string | null,
      id: string,
      interests?: Array< string | null > | null,
      profilePictureUrl?: string | null,
      race?: string | null,
      spokenLanguage?: string | null,
      status?: string | null,
      updatedAt: string,
    } | null,
    receiverId: string,
    senderId: string,
    status?: string | null,
    updatedAt: string,
  } | null,
};

export type GetConnectionRequestQueryVariables = {
  receiverId: string,
  senderId: string,
};

export type GetConnectionRequestQuery = {
  getConnectionRequest?:  {
    __typename: "ConnectionRequest",
    createdAt: string,
    receiverId: string,
    sender?:  {
      __typename: "User",
      aboutMe?: string | null,
      age?: number | null,
      createdAt: string,
      email: string,
      fullName?: string | null,
      gender?: string | null,
      id: string,
      interests?: Array< string | null > | null,
      profilePictureUrl?: string | null,
      race?: string | null,
      spokenLanguage?: string | null,
      status?: string | null,
      updatedAt: string,
    } | null,
    senderId: string,
    status?: string | null,
    updatedAt: string,
  } | null,
};

export type GetDropdownListQueryVariables = {
  id: string,
};

export type GetDropdownListQuery = {
  getDropdownList?:  {
    __typename: "DropdownList",
    createdAt: string,
    id: string,
    name: string,
    options?: Array< string | null > | null,
    updatedAt: string,
  } | null,
};

export type GetMessageQueryVariables = {
  id: string,
};

export type GetMessageQuery = {
  getMessage?:  {
    __typename: "Message",
    content?: string | null,
    createdAt: string,
    id: string,
    room?:  {
      __typename: "Room",
      createdAt: string,
      id: string,
      updatedAt: string,
    } | null,
    roomId: string,
    senderId: string,
    timestamp: string,
    translatedContent?: string | null,
    updatedAt: string,
  } | null,
};

export type GetRaceListQueryVariables = {
  id: string,
};

export type GetRaceListQuery = {
  getRaceList?:  {
    __typename: "RaceList",
    createdAt: string,
    id: string,
    name: string,
    updatedAt: string,
  } | null,
};

export type GetRoomQueryVariables = {
  id: string,
};

export type GetRoomQuery = {
  getRoom?:  {
    __typename: "Room",
    createdAt: string,
    id: string,
    messages?:  {
      __typename: "ModelMessageConnection",
      nextToken?: string | null,
    } | null,
    roomUsers?:  {
      __typename: "ModelRoomUserConnection",
      nextToken?: string | null,
    } | null,
    updatedAt: string,
  } | null,
};

export type GetRoomUserQueryVariables = {
  id: string,
};

export type GetRoomUserQuery = {
  getRoomUser?:  {
    __typename: "RoomUser",
    createdAt: string,
    id: string,
    room?:  {
      __typename: "Room",
      createdAt: string,
      id: string,
      updatedAt: string,
    } | null,
    roomId: string,
    updatedAt: string,
    user?:  {
      __typename: "User",
      aboutMe?: string | null,
      age?: number | null,
      createdAt: string,
      email: string,
      fullName?: string | null,
      gender?: string | null,
      id: string,
      interests?: Array< string | null > | null,
      profilePictureUrl?: string | null,
      race?: string | null,
      spokenLanguage?: string | null,
      status?: string | null,
      updatedAt: string,
    } | null,
    userId: string,
  } | null,
};

export type GetUserQueryVariables = {
  id: string,
};

export type GetUserQuery = {
  getUser?:  {
    __typename: "User",
    aboutMe?: string | null,
    age?: number | null,
    connectionReceived?:  {
      __typename: "ModelConnectionReceivedConnection",
      nextToken?: string | null,
    } | null,
    connectionRequests?:  {
      __typename: "ModelConnectionRequestConnection",
      nextToken?: string | null,
    } | null,
    connections?:  {
      __typename: "ModelConnectionConnection",
      nextToken?: string | null,
    } | null,
    createdAt: string,
    email: string,
    fullName?: string | null,
    gender?: string | null,
    id: string,
    interests?: Array< string | null > | null,
    profilePictureUrl?: string | null,
    race?: string | null,
    rooms?:  {
      __typename: "ModelRoomUserConnection",
      nextToken?: string | null,
    } | null,
    spokenLanguage?: string | null,
    status?: string | null,
    updatedAt: string,
  } | null,
};

export type ListConnectionReceivedsQueryVariables = {
  filter?: ModelConnectionReceivedFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
  receiverId?: ModelIDKeyConditionInput | null,
  senderId?: string | null,
  sortDirection?: ModelSortDirection | null,
};

export type ListConnectionReceivedsQuery = {
  listConnectionReceiveds?:  {
    __typename: "ModelConnectionReceivedConnection",
    items:  Array< {
      __typename: "ConnectionReceived",
      createdAt: string,
      receiverId: string,
      senderId: string,
      status?: string | null,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListConnectionRequestsQueryVariables = {
  filter?: ModelConnectionRequestFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
  receiverId?: ModelIDKeyConditionInput | null,
  senderId?: string | null,
  sortDirection?: ModelSortDirection | null,
};

export type ListConnectionRequestsQuery = {
  listConnectionRequests?:  {
    __typename: "ModelConnectionRequestConnection",
    items:  Array< {
      __typename: "ConnectionRequest",
      createdAt: string,
      receiverId: string,
      senderId: string,
      status?: string | null,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListConnectionsQueryVariables = {
  connectionId?: ModelStringKeyConditionInput | null,
  filter?: ModelConnectionFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
  sortDirection?: ModelSortDirection | null,
  userId?: string | null,
};

export type ListConnectionsQuery = {
  listConnections?:  {
    __typename: "ModelConnectionConnection",
    items:  Array< {
      __typename: "Connection",
      connectionId: string,
      createdAt: string,
      updatedAt: string,
      userId: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListDropdownListsQueryVariables = {
  filter?: ModelDropdownListFilterInput | null,
  id?: string | null,
  limit?: number | null,
  nextToken?: string | null,
  sortDirection?: ModelSortDirection | null,
};

export type ListDropdownListsQuery = {
  listDropdownLists?:  {
    __typename: "ModelDropdownListConnection",
    items:  Array< {
      __typename: "DropdownList",
      createdAt: string,
      id: string,
      name: string,
      options?: Array< string | null > | null,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListMessagesQueryVariables = {
  filter?: ModelMessageFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListMessagesQuery = {
  listMessages?:  {
    __typename: "ModelMessageConnection",
    items:  Array< {
      __typename: "Message",
      content?: string | null,
      createdAt: string,
      id: string,
      roomId: string,
      senderId: string,
      timestamp: string,
      translatedContent?: string | null,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListRaceListsQueryVariables = {
  filter?: ModelRaceListFilterInput | null,
  id?: string | null,
  limit?: number | null,
  nextToken?: string | null,
  sortDirection?: ModelSortDirection | null,
};

export type ListRaceListsQuery = {
  listRaceLists?:  {
    __typename: "ModelRaceListConnection",
    items:  Array< {
      __typename: "RaceList",
      createdAt: string,
      id: string,
      name: string,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListRoomUsersQueryVariables = {
  filter?: ModelRoomUserFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListRoomUsersQuery = {
  listRoomUsers?:  {
    __typename: "ModelRoomUserConnection",
    items:  Array< {
      __typename: "RoomUser",
      createdAt: string,
      id: string,
      roomId: string,
      updatedAt: string,
      userId: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListRoomsQueryVariables = {
  filter?: ModelRoomFilterInput | null,
  id?: string | null,
  limit?: number | null,
  nextToken?: string | null,
  sortDirection?: ModelSortDirection | null,
};

export type ListRoomsQuery = {
  listRooms?:  {
    __typename: "ModelRoomConnection",
    items:  Array< {
      __typename: "Room",
      createdAt: string,
      id: string,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListUsersQueryVariables = {
  filter?: ModelUserFilterInput | null,
  id?: string | null,
  limit?: number | null,
  nextToken?: string | null,
  sortDirection?: ModelSortDirection | null,
};

export type ListUsersQuery = {
  listUsers?:  {
    __typename: "ModelUserConnection",
    items:  Array< {
      __typename: "User",
      aboutMe?: string | null,
      age?: number | null,
      createdAt: string,
      email: string,
      fullName?: string | null,
      gender?: string | null,
      id: string,
      interests?: Array< string | null > | null,
      profilePictureUrl?: string | null,
      race?: string | null,
      spokenLanguage?: string | null,
      status?: string | null,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};